Added colours, began spawn scripting, removed old files
Added the coloured balls (other than red) to their rough correct positions in the scene. Began the process of writing a script to either spawn the balls in their correct positions on scene creation or just to reset position when potted. Some difficulty with game objects on the spots locations as their transforms don't match up to ball transforms.

Removed some old UV maps and extraneous assets. Some still remain and will need to be removed.